PB-114: Moved wrapXCoordinates util
