+++ src/utils/coordinates/__test__/coordinateUtils.spec.js
@@ -9,6 +9,7 @@ import {
     reprojectUnknownSrsCoordsToWGS84,
     toLv95,
 } from '@/utils/coordinates/coordinateUtils'
+import { wrapXCoordinates } from '@/utils/coordinates/coordinateUtils'
 
 describe('Unit test functions from coordinateUtils.js', () => {
     describe('reprojectUnknownSrsCoordsToWGS84(x,y)', () => {
@@ -139,6 +140,37 @@ describe('Unit test functions from coordinateUtils.js', () => {
             result.forEach((coord) => {
                 expect(LV95.isInBounds(...coord)).to.be.true
             })
+        })
+    })
+
+    describe('wrapXCoordinates()', () => {
+        it('Wrap in place', () => {
+            const original = [
+                [300, 300],
+                [360, 360],
+            ]
+            const ref2Original = wrapXCoordinates(original, WGS84, true)
+            expect(ref2Original).to.deep.equal([
+                [-60, 300],
+                [0, 360],
+            ])
+            expect(ref2Original).to.deep.equal(original)
+        })
+        it('Wrap not in place', () => {
+            const original = [
+                [300, 300],
+                [360, 360],
+            ]
+            const ref2Original = wrapXCoordinates(original, WGS84, false)
+            expect(ref2Original).to.deep.equal([
+                [-60, 300],
+                [0, 360],
+            ])
+            expect(ref2Original).to.not.deep.equal(original)
+            expect(original).to.deep.equal([
+                [300, 300],
+                [360, 360],
+            ])
         })
     })
 })
