+++ src/utils/coordinates/coordinateUtils.js
@@ -1,3 +1,5 @@
+import { wrapX } from 'ol/coordinate'
+import { get as getProjection } from 'ol/proj'
 import proj4 from 'proj4'
 
 import log from '../logging'
@@ -127,10 +129,38 @@ export class OutOfBoundsError extends Error {
     }
 }
 
+/**
+ * Convert recursively input coordinates into LV95
+ *
+ * @param {[]} input Coordinate of a point, multipoint or polygon
+ * @param {String} epsg EPSG code of the coordinates
+ * @returns {[]} Reprojected coordinates
+ */
 export function toLv95(input, epsg) {
     if (Array.isArray(input[0])) {
         return input.map((si) => toLv95(si, epsg))
     } else {
         return proj4(epsg, LV95.epsg, [input[0], input[1]])
     }
+}
+
+/**
+ * Wraps the provided coordinates in the world extents (i.e. the coordinate range that if equivalent
+ * to the wgs84 [-180, 180))
+ *
+ * @param {Array} coordinates The coordinates (or array of coordinates) to wrap
+ * @param {CoordinateSystem} projection Projection of the coordinates
+ * @param {boolean} inPlace If false, the original coordinates remain untouched and only a copy is
+ *   modified
+ * @returns If "inPlace", then the same reference as "coords", else a reference to the modified copy
+ */
+export function wrapXCoordinates(coordinates, projection, inPlace = false) {
+    let wrappedCoords = coordinates
+    if (!inPlace) {
+        wrappedCoords = wrappedCoords.slice()
+    }
+    if (Array.isArray(wrappedCoords[0])) {
+        return wrappedCoords.map((c) => wrapXCoordinates(c, projection, inPlace))
+    }
+    return wrapX(wrappedCoords, getProjection(projection.epsg))
 }
